python-brain/requirements.txt
redis>=5.0.0
vaderSentiment>=3.3.2
alpaca-py>=0.25.0
# FinBERT (finance-trained transformer) for smarter buy/sell sentiment
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
accelerate>=0.20.0
# Technical indicators (RSI, etc.) and backtesting
pandas>=2.0.0
finta>=1.3
scikit-learn>=1.3.0
